+++ src/components/pages/aws/regions/index.js
@@ -1,0 +1,3 @@
+import Regions from './regions';
+
+export default Regions;
